moved getColorFromIndex into utils.js
