+++ src/utils/utils.js
@@ -214,3 +214,24 @@ export const copyToClipboard = (copyText, label="Data") => {
             alert(`Failed to copy ${label} to clipboard: ${error}`);
         });
 }
+
+// returns hexadecimal color for interpolated gradient between start and end colors
+//
+// startColor and endColor are rgb triplet integer arrays: [ <r>, <g>, <b> ]
+// steps is how many steps in between first and last color
+// index is which of those step colors to return, 0 based
+export const getColorFromIndex = (index, startColor, endColor, steps) => {
+    function rgbToHex(r, g, b) {
+        const hexR = r.toString(16).padStart(2, "0"); // Convert to hex and pad with 0 if needed
+        const hexG = g.toString(16).padStart(2, "0");
+        const hexB = b.toString(16).padStart(2, "0");
+        return `#${hexR}${hexG}${hexB}`; // Return the hexadecimal color string
+    }
+    if (steps <= 1) {
+        return rgbToHex(endColor[0],endColor[1],endColor[2]);
+    }
+    const r = Math.floor(startColor[0] + (index / (steps -1) * (endColor[0] - startColor[0]) ));
+    const g = Math.floor(startColor[1] + (index / (steps -1) * (endColor[1] - startColor[1]) ));
+    const b = Math.floor(startColor[2] + (index / (steps -1) * (endColor[2] - startColor[2]) ));
+    return rgbToHex(r,g,b);
+}
